Fetch displayname for post's geolocation
We use the Google Maps Geocode API to convert lat/long into names we can
show later on in the client.
To make the tests as independent from the network as possible, vcr is
used to record api responses. From then on it just replays them. Api
responses in the fixtures/vcr_cassettes folder are meant to be checked
in.